+++ test/unit/core-namespace.test.ts
@@ -14,20 +14,4 @@ describe('Core Namespace', () => {
       ).rejects.toThrow();
     });
   });
-
-  describe('isContractAddress()', () => {
-    it('checks if address is a contract', async () => {
-      const contractAddress = '0x00000000006c3852cbEf3e08E8dF289169EdE581'; // Seaport contract address
-      const result = await alchemy.core.isContractAddress(contractAddress);
-
-      expect(result).toBe(true);
-    });
-
-    it('checks if address is an EOA', async () => {
-      const eoaAddress = '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045'; // Vitalik's address
-      const result = await alchemy.core.isContractAddress(eoaAddress);
-
-      expect(result).toBe(false);
-    });
-  });
 });